+++ backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/AuctionController.java
@@ -3,11 +3,11 @@ package com.ddang.ddang.auction.presentation;
 import com.ddang.ddang.auction.application.AuctionService;
 import com.ddang.ddang.auction.application.dto.CreateAuctionDto;
 import com.ddang.ddang.auction.application.dto.ReadAuctionDto;
-import com.ddang.ddang.auction.presentation.dto.CreateAuctionRequest;
-import com.ddang.ddang.auction.presentation.dto.CreateAuctionResponse;
-import com.ddang.ddang.auction.presentation.dto.ReadAuctionDetailResponse;
-import com.ddang.ddang.auction.presentation.dto.ReadAuctionResponse;
-import com.ddang.ddang.auction.presentation.dto.ReadAuctionsResponse;
+import com.ddang.ddang.auction.presentation.dto.request.CreateAuctionRequest;
+import com.ddang.ddang.auction.presentation.dto.response.CreateAuctionResponse;
+import com.ddang.ddang.auction.presentation.dto.response.ReadAuctionDetailResponse;
+import com.ddang.ddang.auction.presentation.dto.response.ReadAuctionResponse;
+import com.ddang.ddang.auction.presentation.dto.response.ReadAuctionsResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
