--- backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/CreateAuctionResponse.java
@@ -1,4 +1,0 @@
-package com.ddang.ddang.auction.presentation.dto;
-
-public record CreateAuctionResponse(Long id) {
-}
