+++ backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/CreateAuctionResponse.java
@@ -1,0 +1,4 @@
+package com.ddang.ddang.auction.presentation.dto.response;
+
+public record CreateAuctionResponse(Long id) {
+}
