+++ backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/DirectRegionResponse.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.auction.presentation.dto;
+package com.ddang.ddang.auction.presentation.dto.response;
 
 // TODO 2차 데모데이 이후 리펙터링 예정
 public record DirectRegionResponse(String first, String second, String third) {
