+++ backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/SellerResponse.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.auction.presentation.dto;
+package com.ddang.ddang.auction.presentation.dto.response;
 
 public record SellerResponse(
         Long id,
