+++ backend/ddang/src/main/java/com/ddang/ddang/region/presentation/dto/response/ReadRegionResponse.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.region.presentation.dto;
+package com.ddang.ddang.region.presentation.dto.response;
 
 import com.ddang.ddang.region.application.dto.ReadRegionDto;
