+++ backend/ddang/src/main/java/com/ddang/ddang/region/presentation/dto/response/ReadRegionsResponse.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.region.presentation.dto;
+package com.ddang.ddang.region.presentation.dto.response;
 
 import java.util.List;
